Nearly finished logic of inserting shoes in shoeListing

+++ app/src/main/java/com/example/shoe_store/model/Shoe.kt
@@ -1,9 +1,9 @@
 package com.example.shoe_store.model
 
 
-data class Shoe (
+data class Shoe(
     var name: String,
-    var size: Double,
+    var size: Int,
     var company: String,
     var description: String,
 )
